More sqlite scaffolding / code reorg
