+++ ibis/impala/client.py
@@ -24,7 +24,7 @@ import ibis.common as com
 
 from ibis.config import options
 from ibis.compat import lzip
-from ibis.client import SQLClient
+from ibis.client import SQLClient, Database, DatabaseEntity
 from ibis.filesystems import HDFS, WebHDFS
 from ibis.impala import udf, ddl
 from ibis.impala.compat import impyla, ImpylaError, HS2Error
@@ -297,7 +297,7 @@ class ImpalaClient(SQLClient):
         database : Database
         """
         # TODO: validate existence of database
-        return Database(name, self)
+        return ImpalaDatabase(name, self)
 
     def list_tables(self, like=None, database=None):
         """
@@ -1188,86 +1188,7 @@ def to_stdout(x):
 # ORM-ish usability layer
 
 
-class Database(object):
-
-    def __init__(self, name, client):
-        self.name = name
-        self.client = client
-
-    def __repr__(self):
-        return "{0}('{1}')".format('Database', self.name)
-
-    def __dir__(self):
-        attrs = dir(type(self))
-        unqualified_tables = [self._unqualify(x) for x in self.tables]
-        return list(sorted(set(attrs + unqualified_tables)))
-
-    def __contains__(self, key):
-        return key in self.tables
-
-    @property
-    def tables(self):
-        return self.list_tables()
-
-    def __getitem__(self, key):
-        return self.table(key)
-
-    def __getattr__(self, key):
-        special_attrs = ['_ipython_display_', 'trait_names',
-                         '_getAttributeNames']
-
-        try:
-            return object.__getattribute__(self, key)
-        except AttributeError:
-            if key in special_attrs:
-                raise
-            return self.table(key)
-
-    def _qualify(self, value):
-        return value
-
-    def _unqualify(self, value):
-        return value
-
-    def drop(self, force=False):
-        """
-        Drop the database
-
-        Parameters
-        ----------
-        drop : boolean, default False
-          Drop any objects if they exist, and do not fail if the databaes does
-          not exist
-        """
-        self.client.drop_database(self.name, force=force)
-
-    def namespace(self, ns):
-        """
-        Creates a derived Database instance for collections of objects having a
-        common prefix. For example, for tables fooa, foob, and fooc, creating
-        the "foo" namespace would enable you to reference those objects as a,
-        b, and c, respectively.
-
-        Returns
-        -------
-        ns : DatabaseNamespace
-        """
-        return DatabaseNamespace(self, ns)
-
-    def table(self, name):
-        """
-        Return a table expression referencing a table in this database
-
-        Returns
-        -------
-        table : TableExpr
-        """
-        qualified_name = self._qualify(name)
-        return self.client.table(qualified_name, self.name)
-
-    def list_tables(self, like=None):
-        return self.client.list_tables(like=self._qualify_like(like),
-                                       database=self.name)
+class ImpalaDatabase(Database):
 
     def list_udfs(self, like=None):
         return self.client.list_udfs(like=self._qualify_like(like),
@@ -1276,19 +1197,6 @@ class Database(object):
     def list_udas(self, like=None):
         return self.client.list_udas(like=self._qualify_like(like),
                                      database=self.name)
-
-    def _qualify_like(self, like):
-        return like
-
-
-class DatabaseEntity(object):
-    pass
-
-
-class View(DatabaseEntity):
-
-    def drop(self):
-        pass
 
 
 class ScalarFunction(DatabaseEntity):
@@ -1301,37 +1209,6 @@ class AggregateFunction(DatabaseEntity):
 
     def drop(self):
         pass
-
-
-class DatabaseNamespace(Database):
-
-    def __init__(self, parent, namespace):
-        self.parent = parent
-        self.namespace = namespace
-
-    def __repr__(self):
-        return ("{0}(database={1!r}, namespace={2!r})"
-                .format('DatabaseNamespace', self.name, self.namespace))
-
-    @property
-    def client(self):
-        return self.parent.client
-
-    @property
-    def name(self):
-        return self.parent.name
-
-    def _qualify(self, value):
-        return self.namespace + value
-
-    def _unqualify(self, value):
-        return value.replace(self.namespace, '', 1)
-
-    def _qualify_like(self, like):
-        if like:
-            return self.namespace + like
-        else:
-            return '{0}*'.format(self.namespace)
 
 
 class ImpalaTable(ir.TableExpr, DatabaseEntity):
